common and unique in the same form

+++ ERP_Cosmus/product/views.py
@@ -544,31 +544,46 @@ def product2subcategoryajax(request):
 
 
 def product2item(request,product_refrence_id):
+    print(request.POST)
     items = Item_Creation.objects.all()
     product_refrence_no = product_refrence_id
     Products_all = PProduct_Creation.objects.filter(Product__Product_Refrence_ID=product_refrence_id)
 
     product2item_instances = product_2_item_through_table.objects.filter(PProduct_pk__Product__Product_Refrence_ID=product_refrence_id, common_unique = False)
-
     formset_single = Product2ItemFormset(queryset=product2item_instances , prefix='product2itemuniqueformset')
+
+    product2item_common_instances = product_2_item_through_table.objects.filter(PProduct_pk__Product__Product_Refrence_ID=product_refrence_id, common_unique = True)
+    print('product2item_common_instances',product2item_common_instances)
+    formset_common = Product2CommonItemFormSet(queryset=product2item_common_instances,prefix='product2itemcommonformset')
+
 
     if request.method == 'POST':
         formset_single = Product2ItemFormset(request.POST, queryset=product2item_instances, prefix='product2itemuniqueformset')
-        print(request.POST)
-        if formset_single.is_valid():
+        formset_common = Product2CommonItemFormSet(request.POST, queryset=product2item_common_instances, prefix='product2itemcommonformset')
+        
+        if formset_single.is_valid() and formset_common.is_valid():
             # when using form.save(commit=False) we need to  explicitly delete forms marked in has_deleted
             for form in formset_single.deleted_forms:
                 if form.instance.pk:  # Ensure the form instance has a primary key before attempting deletion
                     form.instance.delete()
             
             for form in formset_single:
-                print('form.cleaned_data1',form.cleaned_data.get('DELETE'))
                 if not form.cleaned_data.get('DELETE'): # check if form not in deleted forms to avoid saving it again 
-                    print('form.cleaned_data1',form.cleaned_data.get('Item_pk'))
                     if form.cleaned_data.get('Item_pk'):  # Check if the form has 'Item_pk' filled
-                        print('form.cleaned_data',form.cleaned_data)
                         p2i_instance = form.save(commit = False)
                         p2i_instance.common_unique = False
+                        p2i_instance.save()
+            
+            #common formset
+            for form in formset_common.deleted_forms:
+                if form.instance.pk:  # Ensure the form instance has a primary key before attempting deletion
+                    form.instance.delete()
+            
+            for form in formset_common:
+                if not form.cleaned_data.get('DELETE'): # check if form not in deleted forms to avoid saving it again 
+                    if form.cleaned_data.get('Item_pk'):  # Check if the form has 'Item_pk' filled
+                        p2i_instance = form.save(commit = False)
+                        p2i_instance.common_unique = True
                         p2i_instance.save()
 
 
@@ -586,6 +601,11 @@ def product2item(request,product_refrence_id):
     return render(request, 'production/product2itemsetproduction.html', { 'formset_single':formset_single,
                                                                'Products_all':Products_all,
                                                                'items':items,'product_refrence_no': product_refrence_no})
+
+
+
+
+
 def product2commonitem(request,product_id): 
     items = Item_Creation.objects.all()
     product_refrence_no = product_id
